move things into parquet
